first pass at rearranging the directory structure
