+++ db-outputs/PureFusionMaterials_mcnpAtomfrac.txt
