+++ db-outputs/PureFusionMaterials_mcnpMassfrac.txt
