+++ examples/mixPureFusionMaterials.py
@@ -1,56 +1,9 @@
-#! /usr/bin/python
-#
-# -updated for python3 (print) and updated for changes in python modules
-#
-#
-# mixes pure fusion materials based on FESS-FNSF, ARIES, EUDEMO and other
-# designs
-# -can be used for mixing homogenized regions
-#
-#
-#
+
 import os
-from pyne import material
-from pyne.material import Material, MultiMaterial
-from pyne.material_library import MaterialLibrary
 
-
-def get_consituent_citations(materials):
-    citation_str = ""
-    for mat in materials:
-        citation_str = " ".join([citation_str, mat.metadata["citation"]])
-    return citation_str
-
-
-# Mix Materials by Volume
-def mix_by_volume(material_library, vol_fracs, citation):
-    """
-    Mixes materials by volume, adds list of constituent citations to the
-    metadata
-
-    Arguments:
-        material_library (PyNE material library): library containing constituent
-            materials.
-        vol_fracs (dict): dictionary where the keys are names of materials (str)
-            and values are the volume fraction (float)
-        citation (str): citation for the mixture
-    """
-    mix_dict = {}
-
-    for name, volume_fraction in vol_fracs.items():
-        mix_dict[material_library[name]] = volume_fraction
-
-    mix = MultiMaterial(mix_dict)
-    mat = mix.mix_by_volume()
-    mat.metadata["mixture_citation"] = citation
-    mat.metadata["constituent_citation"] = get_consituent_citations(
-        list(mix_dict.keys())
-    )
-    return mat
-
+import material_db_tools as mdbt
 
 mat_data = {}
-
 
 # fullreference: DavisFusEngDes_2018
 # https://doi.org/10.1016/j.fusengdes.2017.06.008
@@ -221,7 +174,7 @@ def main():
     # create material library object
     mixmat_lib = MaterialLibrary()
     for mat_name, mat_input in mat_data.items():
-        mixmat_lib[mat_name] = mix_by_volume(
+        mixmat_lib[mat_name] = mdbt.mix_by_volume(
             mat_lib, mat_input["vol_fracs"], mat_input["mixture_citation"]
         )
